refactor(db): one place to declare a column, and errors that surface
The schema was two lists that had to agree: `SCHEMA` for a fresh database
and 52 append-only `ALTER`/`CREATE` statements for an existing one, applied
under `let _ = conn.execute(sql, [])`. Every error was discarded, so a
statement that failed for a real reason was indistinguishable from one that
had already done its work, and the ordering the comments kept apologising for
was enforced by nothing.

The declared DDL is now the only source of truth. On open, a reconciler
creates missing tables, adds missing columns, then creates missing indexes --
three passes, because an index names columns. Adding a column is one edit to
one file; existing databases pick it up on the next boot.

Deriving the true schema mechanically (declared vs declared-plus-migrations)
found 13 objects that existed only because a swallowed statement happened to
work: users.pin_hash, users.language, users.audio_language,
users.subtitle_language, users.last_seen, access_tokens.user_agent,
sessions.access_token, the passkeys table and five indexes. They are declared
now.

`ddl.rs` was 624 lines; it is split by domain, and the eight tables a module
owns are grouped in `module_owned.rs` rather than spread through files that
imply the core owns those nouns. Nothing moved into or out of the core
database -- the grouping only counts the debt in one place.

`media/shows.rs` reads its rows by column name instead of by index, so a
column added to the SELECT can no longer shift what the mapper sees.

Behaviour change: a database is no longer upgraded by replaying history, so
the drops, renames and backfills the old list carried do not run. What the
reconciler cannot express -- a drop, a rename, a backfill, a tightened
constraint -- is invisible to it and still needs a statement of its own.
