+++ server/crates/kroma-db/src/schema.rs
@@ -1,4 +1,4 @@
-//! Opening a database: the pool, the schema and the migrations that follow it.
+//! Opening a database: the pool, and the declared schema it is brought up to.
 
 use std::path::Path;
 use std::sync::{Arc, Mutex};
@@ -9,10 +9,10 @@ use rusqlite::Connection;
 use super::{Pool, PoolInner};
 
 mod ddl;
-mod migrations;
+pub(crate) mod declared;
+mod reconcile;
 
 pub(crate) use ddl::{FILE_COLS, ITEM_COLS, PRAGMAS, SCHEMA};
-pub(crate) use migrations::MIGRATIONS;
 
 /// A pool over `path` with the pragmas applied and NO schema.
 ///
@@ -49,16 +49,8 @@ pub(crate) fn pool_at(
 pub fn init(path: &Path) -> Result<Pool> {
     let pool = pool_at(path, 8, None)?;
     let conn = pool.get()?;
-    conn.execute_batch(SCHEMA)
-        .context("failed to apply schema")?;
-    migrate(&conn);
+    reconcile::apply(&conn)?;
     Ok(pool)
-}
-
-fn migrate(conn: &Connection) {
-    for sql in MIGRATIONS {
-        let _ = conn.execute(sql, []);
-    }
 }
 
 /// Apply a module's own schema after the core schema at DB init (see
@@ -107,96 +99,6 @@ mod tests {
         assert!(
             format!("{err:#}").contains("failed to apply module schema"),
             "{err:#}"
-        );
-    }
-
-    #[test]
-    fn init_opens_a_database_whose_wanted_table_predates_the_backoff_columns() {
-        let dir = kroma_testing::temp_dir("schema-wanted-upgrade");
-        let path = dir.path().join("kroma.db");
-
-        let old = Connection::open(&path).unwrap();
-        old.execute_batch(
-            "CREATE TABLE wanted (\
-                 id             TEXT PRIMARY KEY,\
-                 request_id     TEXT NOT NULL,\
-                 kind           TEXT NOT NULL,\
-                 tmdb_id        INTEGER NOT NULL,\
-                 title          TEXT NOT NULL,\
-                 season         INTEGER,\
-                 episode        INTEGER,\
-                 air_date       TEXT,\
-                 status         TEXT NOT NULL DEFAULT 'wanted',\
-                 last_search_at INTEGER,\
-                 updated_at     INTEGER NOT NULL);\
-             CREATE INDEX idx_wanted_search ON wanted(status, last_search_at);\
-             INSERT INTO wanted (id, request_id, kind, tmdb_id, title, updated_at)\
-                 VALUES ('w1', 'r1', 'episode', 42, 'Show', 0);",
-        )
-        .unwrap();
-        drop(old);
-
-        let pool = init(&path).expect("an existing database must still open");
-        let conn = pool.get().unwrap();
-
-        let due: i64 = conn
-            .query_row(
-                "SELECT COUNT(*) FROM wanted WHERE next_search_at IS NULL",
-                [],
-                |r| r.get(0),
-            )
-            .unwrap();
-        assert_eq!(due, 1, "rows from before the column keep their turn");
-
-        let indexes: Vec<String> = conn
-            .prepare("SELECT name FROM sqlite_master WHERE type = 'index' AND tbl_name = 'wanted'")
-            .unwrap()
-            .query_map([], |r| r.get(0))
-            .unwrap()
-            .collect::<std::result::Result<_, _>>()
-            .unwrap();
-        assert!(indexes.iter().any(|n| n == "idx_wanted_due"), "{indexes:?}");
-        assert!(
-            !indexes.iter().any(|n| n == "idx_wanted_search"),
-            "{indexes:?}"
-        );
-    }
-
-    #[test]
-    fn init_repins_downloads_a_previous_release_marked_manual() {
-        let dir = kroma_testing::temp_dir("schema-match-source");
-        let path = dir.path().join("kroma.db");
-        let old = init(&path).unwrap();
-        old.get()
-            .unwrap()
-            .execute_batch(
-                "INSERT INTO downloads (id, client_id, client_ref, kind, tmdb_id, release_title, \
-                     magnet_or_url, grabbed_at, match_source) \
-                 VALUES ('d1', 'embedded', '', 'movie', 603, 'R', 'm', 0, 'manual'), \
-                        ('d2', 'embedded', '', 'movie', 0, 'R', 'm', 0, 'none'), \
-                        ('d3', 'embedded', '', 'movie', 42, 'R', 'm', 0, 'auto');",
-            )
-            .unwrap();
-        drop(old);
-
-        let pool = init(&path).unwrap();
-
-        let sources: Vec<Option<String>> = pool
-            .get()
-            .unwrap()
-            .prepare("SELECT match_source FROM downloads ORDER BY id")
-            .unwrap()
-            .query_map([], |r| r.get(0))
-            .unwrap()
-            .collect::<std::result::Result<_, _>>()
-            .unwrap();
-        assert_eq!(
-            sources,
-            [
-                Some("pinned".to_string()),
-                Some("none".to_string()),
-                Some("auto".to_string())
-            ]
         );
     }
 }
