+++ server/crates/kroma-db/src/schema/ddl/metadata.rs
@@ -1,0 +1,89 @@
+//! What a title IS, apart from the language it is described in.
+
+pub(super) const SCHEMA: &str = "
+    -- Global editorial collections, regenerated by the `sections.curate` job
+    -- (member ids resolved at write time). Localized title/reason live in
+    -- `translations` (`subject_kind='curated'`), not in per-language columns.
+    CREATE TABLE IF NOT EXISTS curated_sections (
+        key        TEXT PRIMARY KEY,
+        rank       INTEGER NOT NULL DEFAULT 0,
+        source     TEXT NOT NULL DEFAULT 'llm',
+        item_ids   TEXT NOT NULL DEFAULT '[]',
+        updated_at INTEGER NOT NULL
+    );
+
+    -- Per-item AI suggestions, lazily generated on first view; empty item_ids is
+    -- a terminal `tried, nothing usable` marker. Localized reason lives in
+    -- `translations` (`subject_kind='suggestion'`), not in per-language columns.
+    CREATE TABLE IF NOT EXISTS item_suggestions (
+        item_id    TEXT PRIMARY KEY,
+        item_ids   TEXT NOT NULL DEFAULT '[]',
+        updated_at INTEGER NOT NULL
+    );
+
+    -- Availability matching is a seek on `metadata_core.tmdb_id` (a real indexed
+    -- column, see idx_meta_core_tmdb below); the old json_extract expression
+    -- indexes on the metadata blob are retired in `migrate`.
+
+    -- An operator-chosen TMDB id for one catalog subject, set from the 'fix the
+    -- match' picker. Enrichment consults this BEFORE any title guess, so a
+    -- correction is authoritative and survives every re-scan. Distinct from
+    -- `acq_file_tmdb` above (set automatically at import): this operator choice
+    -- takes precedence over it.
+    CREATE TABLE IF NOT EXISTS tmdb_pin (
+        subject_kind TEXT NOT NULL,          -- 'item' | 'show'
+        subject_id   TEXT NOT NULL,
+        tmdb_id      INTEGER NOT NULL,
+        updated_at   INTEGER NOT NULL,
+        PRIMARY KEY (subject_kind, subject_id)
+    );
+
+    -- Language-INVARIANT resolved metadata, one row per catalog subject. Split
+    -- out of the per-item `metadata` JSON so identity/availability/art/cast don't
+    -- depend on which language was fetched, and switching UI language never
+    -- touches this row (nor the embeddings derived from it).
+    --
+    -- A provider id is stored ON THE RECORD IT IDENTIFIES, under a
+    -- provider-qualified name: this title's own ids are the columns below, and
+    -- a credited person's rides on its member inside cast_json / crew_json (a
+    -- serde field, so people cost no schema change). Only a bare reference list
+    -- the title holds, like genres, earns a column of its own here.
+    CREATE TABLE IF NOT EXISTS metadata_core (
+        subject_kind TEXT NOT NULL,          -- 'item' | 'show'
+        subject_id   TEXT NOT NULL,
+        tmdb_id      INTEGER,
+        imdb_id      TEXT,
+        tvdb_id      INTEGER,
+        release_date TEXT,
+        rating       REAL,
+        poster_url   TEXT,
+        backdrop_url TEXT,
+        logo_url     TEXT,
+        cast_json    TEXT NOT NULL DEFAULT '[]',
+        crew_json    TEXT NOT NULL DEFAULT '[]',
+        -- The NAMES are localized and live in `translations`; these do not.
+        tmdb_genre_ids TEXT NOT NULL DEFAULT '[]',
+        updated_at   INTEGER NOT NULL,
+        PRIMARY KEY (subject_kind, subject_id)
+    );
+    CREATE INDEX IF NOT EXISTS idx_meta_core_tmdb ON metadata_core(tmdb_id);
+
+    -- The GENERIC per-language translation cache: one table for every localized
+    -- string in the app, so adding a language is inserting rows, never a schema
+    -- change. `resolve` falls back requested lang -> en, and skips a row that
+    -- carries only a revision: that is a marker saying the language was asked
+    -- for and had nothing, not something to serve.
+    CREATE TABLE IF NOT EXISTS translations (
+        subject_kind TEXT NOT NULL,   -- 'item'|'show'|'episode'|'season_cast'|'curated'|'suggestion'
+        subject_id   TEXT NOT NULL,
+        lang         TEXT NOT NULL,   -- a code from i18n::SUPPORTED_LOCALES
+        source       TEXT NOT NULL,   -- 'tmdb' | 'llm' | 'manual'
+        data         TEXT NOT NULL,   -- JSON: {title,overview,tagline,genres,posterUrl?,logoUrl?,characters?,reason?,rev?}
+        updated_at   INTEGER NOT NULL,
+        PRIMARY KEY (subject_kind, subject_id, lang)
+    );
+    -- For the reads that name a language but no subject: which genres the
+    -- catalogue holds, and which titles carry one of them. A read that does
+    -- name its subjects is served by the primary key instead.
+    CREATE INDEX IF NOT EXISTS idx_translations_lang ON translations(subject_kind, lang);
+";
